Ignore null groupId when restoring default account.
When restoring accounts in sPhoneAccountXml.readFromXml, the groupId of
the newly created accounts isn't filled (setGroupId isn't called on the
builder) and thus none of the accounts have a group ID assigned. When
then running setUserSelectedOutgoingPhoneAccount(), the used groupId
consequently is empty, thus it doesn't seem correct to mandate it being
non-empty when later restoring it.

Change-Id: I14256616bae62e6689e0d2d4751f41bd90dfcba6

Signed-off-by: xyyx <xyyx@mail.ru>